added cutting wedge to extrude shapes and tf coils
