+++ paramak/parametric_components/toroidal_field_coil_princeton_d.py
@@ -6,6 +6,7 @@ from scipy import integrate
 from scipy.optimize import minimize
 
 from paramak import ExtrudeMixedShape
+from paramak.utils import calculate_wedge_cut
 
 
 class ToroidalFieldCoilPrincetonD(ExtrudeMixedShape):
@@ -37,6 +38,7 @@ class ToroidalFieldCoilPrincetonD(ExtrudeMixedShape):
         thickness,
         distance,
         number_of_coils,
+        rotation_angle=360,
         vertical_displacement=0.0,
         with_inner_leg=True,
         stp_filename="ToroidalFieldCoilPrincetonD.stp",
@@ -60,6 +62,7 @@ class ToroidalFieldCoilPrincetonD(ExtrudeMixedShape):
         self.number_of_coils = number_of_coils
         self.vertical_displacement = vertical_displacement
         self.with_inner_leg = with_inner_leg
+        self.rotation_angle = rotation_angle
 
     @property
     def azimuth_placement_angle(self):
@@ -69,6 +72,14 @@ class ToroidalFieldCoilPrincetonD(ExtrudeMixedShape):
     @azimuth_placement_angle.setter
     def azimuth_placement_angle(self, value):
         self._azimuth_placement_angle = value
+
+    @property
+    def rotation_angle(self):
+        return self._rotation_angle
+
+    @rotation_angle.setter
+    def rotation_angle(self, value):
+        self._rotation_angle = value
 
     def _compute_inner_points(self, R1, R2):
         """Computes the inner curve points
