+++ paramak/parametric_shapes/extruded_mixed_shape.py
@@ -3,6 +3,7 @@ from collections import Iterable
 import cadquery as cq
 
 from paramak import Shape
+from paramak.utils import calculate_wedge_cut
 
 
 class ExtrudeMixedShape(Shape):
@@ -20,6 +21,7 @@ class ExtrudeMixedShape(Shape):
     def __init__(
         self,
         distance,
+        rotation_angle=360,
         stp_filename="ExtrudeMixedShape.stp",
         stl_filename="ExtrudeMixedShape.stl",
         **kwargs
@@ -32,6 +34,7 @@ class ExtrudeMixedShape(Shape):
         )
 
         self.distance = distance
+        self.rotation_angle = rotation_angle
 
     @property
     def distance(self):
@@ -40,6 +43,14 @@ class ExtrudeMixedShape(Shape):
     @distance.setter
     def distance(self, value):
         self._distance = value
+
+    @property
+    def rotation_angle(self):
+        return self._rotation_angle
+
+    @rotation_angle.setter
+    def rotation_angle(self, value):
+        self._rotation_angle = value
 
     def create_solid(self):
         """Creates an extruded 3d solid using points connected with straight
@@ -108,6 +119,7 @@ class ExtrudeMixedShape(Shape):
             solid = solid.rotate(
                 (0, 0, -1), (0, 0, 1), self.azimuth_placement_angle)
 
+        calculate_wedge_cut(self)
         self.perform_boolean_operations(solid)
 
         return solid
