+++ paramak/parametric_shapes/extruded_spline_shape.py
@@ -3,6 +3,7 @@ from collections import Iterable
 import cadquery as cq
 
 from paramak import Shape
+from paramak.utils import calculate_wedge_cut
 
 
 class ExtrudeSplineShape(Shape):
@@ -19,6 +20,7 @@ class ExtrudeSplineShape(Shape):
     def __init__(
         self,
         distance,
+        rotation_angle=360,
         stp_filename="ExtrudeSplineShape.stp",
         stl_filename="ExtrudeSplineShape.stl",
         **kwargs
@@ -31,6 +33,7 @@ class ExtrudeSplineShape(Shape):
         )
 
         self.distance = distance
+        self.rotation_angle = rotation_angle
 
     @property
     def distance(self):
@@ -39,6 +42,14 @@ class ExtrudeSplineShape(Shape):
     @distance.setter
     def distance(self, value):
         self._distance = value
+
+    @property
+    def rotation_angle(self):
+        return self._rotation_angle
+
+    @rotation_angle.setter
+    def rotation_angle(self, value):
+        self._rotation_angle = value
 
     def create_solid(self):
         """Creates an extruded 3d solid using points connected with spline
@@ -74,6 +85,7 @@ class ExtrudeSplineShape(Shape):
             solid = solid.rotate(
                 (0, 0, -1), (0, 0, 1), self.azimuth_placement_angle)
 
+        calculate_wedge_cut(self)
         self.perform_boolean_operations(solid)
 
         return solid
